[WIP] create tf.function and update to tensorflow 2.0

+++ keras_frcnn/roi_helpers.py
@@ -93,7 +93,7 @@ def calc_iou(R, img_data, C, class_mapping):
 	Y1 = np.array(y_class_num)
 	Y2 = np.concatenate([np.array(y_class_regr_label),np.array(y_class_regr_coords)],axis=1)
 
-	return np.expand_dims(X, axis=0), np.expand_dims(Y1, axis=0), np.expand_dims(Y2, axis=0), IoUs
+	return np.expand_dims(X, axis=0).astype('float32'), np.expand_dims(Y1, axis=0).astype('float32'), np.expand_dims(Y2, axis=0).astype('float32'), IoUs
 
 def apply_regr(x, y, w, h, tx, ty, tw, th):
 	try:
